Move class to tests
